fix: in actuarius, the last character of an value which is not semicolon-terminated, e.g. 1, is not read
fix: in actuarius, the last character of an value which is not semicolon-terminated, e.g. 2, is not read

fix: in actuarius, the last character of an value which is not semicolon-terminated, e.g. 3, is not read

fix: in actuarius, the last character of an value which is not semicolon-terminated, e.g. 45, is not read